requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
gunicorn==20.1.0